+++ js/nomnoml.js
@@ -326,7 +326,7 @@ skanaar.Svg = function (globalStyle){
 		fillText: function (text, x, y){
 			if (lastDefined('textAlign') === 'center')
 				x -= this.measureText(text).width/2
-			return newElement('text', { x: tX(x), y: tY(y) }, text)
+			return newElement('text', { x: tX(x), y: tY(y) }, _.escape(text))
 		},
 		lineCap: function (cap){ globalStyle += ';stroke-linecap:'+cap },
 		lineJoin: function (join){ globalStyle += ';stroke-linejoin:'+join },
@@ -335,7 +335,12 @@ skanaar.Svg = function (globalStyle){
 		},
 		lineWidth: function (w){ globalStyle += ';stroke-width:'+w},
 		measureText: function (s){
-			return { width: s.length * 8.5 }
+			return {
+				width: skanaar.sum(s, function (c){
+					if (c === 'M' || c === 'W') { return 14 }
+					return c.charCodeAt(0) < 200 ? 9.5 : 16
+				})
+			}
 		},
 		moveTo: function (x, y){
 			last(elements).attr.d += ('M' + tX(x) + ' ' + tY(y) + ' ')
@@ -364,6 +369,9 @@ skanaar.Svg = function (globalStyle){
       attrs.baseProfile = attrs.baseProfile || 'full';
       attrs.width = attrs.width || '100%';
       attrs.height = attrs.height || '100%';
+      if(attrs.width !== '100%' && attrs.height != '100%') {
+        attrs.viewbox = '0 0 ' + attrs.width + ' ' + attrs.height;
+	  }
       attrs.xmlns = attrs.xmlns || 'http://www.w3.org/2000/svg';
       attrs['xmlns:xlink'] = attrs['xmlns:xlink'] || 'http://www.w3.org/1999/xlink';
       attrs['xmlns:ev']  = attrs['xmlns:ev'] || 'http://www.w3.org/2001/xml-events';
@@ -463,7 +471,7 @@ symbols_: {"error":2,"root":3,"compartment":4,"EOF":5,"slot":6,"IDENT":7,"class"
 terminals_: {2:"error",5:"EOF",7:"IDENT",10:"SEP",12:"|",13:"[",14:"]"},
 productions_: [0,[3,2],[6,1],[6,1],[6,1],[4,1],[4,3],[11,1],[11,3],[11,2],[9,3],[8,3]],
 performAction: function anonymous(yytext, yyleng, yylineno, yy, yystate /* action[1] */, $$ /* vstack */, _$ /* lstack */
-/**/) {
+/*``*/) {
 /* this == yyval */
 
 var $0 = $$.length - 1;
@@ -969,7 +977,7 @@ stateStackSize:function stateStackSize() {
     },
 options: {},
 performAction: function anonymous(yy,yy_,$avoiding_name_collisions,YY_START
-/**/) {
+/*``*/) {
 
 var YYSTATE=YY_START;
 switch($avoiding_name_collisions) {
@@ -1120,6 +1128,7 @@ nomnoml.styles = {
   SENDER:   { center: 0, bold: 0, underline: 0, italic: 0, dashed: 0, empty: 0, hull: 'auto', visual: 'sender' },
   START:    { center: 1, bold: 0, underline: 0, italic: 0, dashed: 0, empty: 1, hull: 'icon', visual: 'start' },
   STATE:    { center: 1, bold: 0, underline: 0, italic: 0, dashed: 0, empty: 0, hull: 'auto', visual: 'roundrect' },
+  TRANSCEIVER:{ center: 0, bold: 0, underline: 0, italic: 0, dashed: 0, empty: 0, hull: 'auto', visual: 'transceiver' },
   USECASE:  { center: 1, bold: 0, underline: 0, italic: 0, dashed: 0, empty: 0, hull: 'auto', visual: 'ellipse' },
 }
 
@@ -1200,13 +1209,13 @@ nomnoml.visualizers = {
     ]).fillAndStroke()
   },
   receiver : function (node, x, y, padding, config, g) {
-    g.circuit([
-      {x: x, y: y},
-      {x: x+node.width+padding, y: y},
-      {x: x+node.width-padding, y: y+node.height/2},
-      {x: x+node.width+padding, y: y+node.height},
-      {x: x, y: y+node.height}
-    ]).fillAndStroke()
+      g.circuit([
+        {x: x-padding, y: y},
+        {x: x+node.width, y: y},
+        {x: x+node.width, y: y+node.height},
+        {x: x-padding, y: y+node.height},
+        {x: x, y: y+node.height/2},
+      ]).fillAndStroke()
   },
   rhomb : function (node, x, y, padding, config, g) {
     g.circuit([
@@ -1221,17 +1230,27 @@ nomnoml.visualizers = {
     g.roundRect(x, y, node.width, node.height, r).fillAndStroke()
   },
   sender : function (node, x, y, padding, config, g) {
-    g.circuit([
-      {x: x, y: y},
-      {x: x+node.width-padding, y: y},
-      {x: x+node.width+padding, y: y+node.height/2},
-      {x: x+node.width-padding, y: y+node.height},
-      {x: x, y: y+node.height}
-    ]).fillAndStroke()
+      g.circuit([
+        {x: x, y: y},
+        {x: x+node.width-padding, y: y},
+        {x: x+node.width, y: y+node.height/2},
+        {x: x+node.width-padding, y: y+node.height},
+        {x: x, y: y+node.height}
+      ]).fillAndStroke()
   },
   start : function (node, x, y, padding, config, g) {
     g.fillStyle(config.stroke)
     g.circle(node.x, y+node.height/2, node.height/2.5).fill()
+  },
+  transceiver : function (node, x, y, padding, config, g) {
+      g.circuit([
+        {x: x-padding, y: y},
+        {x: x+node.width, y: y},
+        {x: x+node.width+padding, y: y+node.height/2},
+        {x: x+node.width, y: y+node.height},
+        {x: x-padding, y: y+node.height},
+        {x: x, y: y+node.height/2}
+      ]).fillAndStroke()
   },
 };
 var nomnoml = nomnoml || {}
@@ -1318,12 +1337,15 @@ nomnoml.layout = function (measurer, config, ast){
 			clas.height = 0
 			return
 		}
+		var oldDir = config.direction;
+		config.direction = style.direction || config.direction;
 		_.each(clas.compartments, layoutCompartment)
 		clas.width = _.max(_.pluck(clas.compartments, 'width'))
 		clas.height = skanaar.sum(clas.compartments, 'height')
 		clas.x = clas.width/2
 		clas.y = clas.height/2
 		_.each(clas.compartments, function(co){ co.width = clas.width })
+		config.direction = oldDir;
 	}
 	layoutCompartment(ast)
 	return ast
@@ -1417,10 +1439,58 @@ nomnoml.render = function (graphics, config, compartment, setFont){
 
 	var empty = false, filled = true, diamond = true
 
+    function renderLabel(text, refPoint, quadrant){
+		if (text) {
+			var fontSize = config.fontSize
+			var lines = text.split("`")
+			var area = {
+				width : _.max(_.map(lines, function(l){ return g.measureText(l).width })),
+				height : fontSize*lines.length
+			}
+			var origin = {
+				x: (quadrant === 1) || (quadrant === 4) ? refPoint.x + padding : refPoint.x - area.width - padding,
+				y: (quadrant === 3) || (quadrant === 4) ? refPoint.y + padding : refPoint.y - area.height - padding
+			}
+			_.each(lines, function(l, i){ g.fillText(l, origin.x, origin.y + fontSize*(i+1)) })
+		}
+	}
+
+	// find basic quadrant using relative position of endpoint and block rectangle
+	function findLabelQuadrant(point, rect, def) {
+		if (point.x < rect.x && point.y < rect.y-rect.height/2) return 1;
+		if (point.y > rect.y && point.x > rect.x+rect.width/2) return 1;
+		
+		if (point.x > rect.x && point.y < rect.y-rect.height/2) return 2;
+		if (point.y > rect.y && point.x < rect.x-rect.width/2) return 2;
+
+		if (point.x > rect.x && point.y > rect.y+rect.height/2) return 3;
+		if (point.y < rect.y && point.x < rect.x-rect.width/2) return 3;
+
+		if (point.x < rect.x && point.y > rect.y+rect.height/2) return 4;
+		if (point.y < rect.y && point.x > rect.x+rect.width/2) return 4;
+
+		return def;
+	}
+
+	// Flip basic label quadrant if needed, to avoid crossing a bent relationship line
+	function adjustLabelQuadrant(quadrant, point, opposite) {
+		if ((opposite.x == point.x) || (opposite.y == point.y)) return quadrant;
+		var flipHorizontally = [4, 3, 2, 1]
+		var flipVertically = [2, 1, 4, 3]
+		var oppositeQuadrant = (opposite.y < point.y) ?
+							((opposite.x < point.x) ? 2 : 1) :
+							((opposite.x < point.x) ? 3 : 4);
+		// if an opposite relation end is in the same quadrant as a label, we need to flip the label
+		if (oppositeQuadrant === quadrant) {
+			if (config.direction === "LR") return flipHorizontally[quadrant-1];
+			if (config.direction === "TD") return flipVertically[quadrant-1];
+		}
+		return quadrant; 	
+	}
+
 	function renderRelation(r, compartment){
 		var startNode = _.findWhere(compartment.nodes, {name:r.start})
 		var endNode = _.findWhere(compartment.nodes, {name:r.end})
-
 		var start = rectIntersection(r.path[1], _.first(r.path), startNode)
 		var end = rectIntersection(r.path[r.path.length-2], _.last(r.path), endNode)
 
@@ -1429,10 +1499,9 @@ nomnoml.render = function (graphics, config, compartment, setFont){
 
 		g.fillStyle(config.stroke)
 		setFont(config, 'normal')
-		var textW = g.measureText(r.endLabel).width
-		var labelX = config.direction === 'LR' ? -padding-textW : padding
-		if (r.startLabel) g.fillText(r.startLabel, start.x+padding, start.y+padding+fontSize)
-		if (r.endLabel)	 g.fillText(r.endLabel, end.x+labelX, end.y-padding)
+
+		renderLabel(r.startLabel, start, adjustLabelQuadrant(findLabelQuadrant(start, startNode, 4), start, end))
+		renderLabel(r.endLabel, end, adjustLabelQuadrant(findLabelQuadrant(end, endNode, 2), end, start))
 
 		if (r.assoc !== '-/-'){
 			if (g.setLineDash && skanaar.hasSubstring(r.assoc, '--')){
@@ -1531,7 +1600,8 @@ var nomnoml = nomnoml || {};
 				dashed: _.contains(styleDef, 'dashed'),
 				empty: _.contains(styleDef, 'empty'),
 				fill: _.last(styleDef.match('fill=([^ ]*)')),
-				visual: _.last(styleDef.match('visual=([^ ]*)')) || 'class'
+				visual: _.last(styleDef.match('visual=([^ ]*)')) || 'class',
+				direction: { down: 'TB', right: 'LR' }[_.last(styleDef.match('direction=([^ ]*)'))] || 'TB'
 			}
 		})
 		return {