+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# v1.0.1
+## 04-07-2018
+
+1. [](#bugfix)
+    * Script-execution timing
+2. [](#new)
+    * Update scripts to [dev-master](344e3f58781c2185a4087cd9680929ec9ced358b)
+    * Default-options array
+3. [](#improved)
+    * README
+    * Code quality
+
 # v1.0.0
 ## 23-06-2017